+++ src/Formatter/DefaultConfig.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Inspirum\XML\Formatter;
+
+final readonly class DefaultConfig extends BaseConfig
+{
+}