+++ CHANGELOG.md
@@ -6,7 +6,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased](https://github.com/inspirum/xml-php/compare/v2.3.1...master)
+## [Unreleased](https://github.com/inspirum/xml-php/compare/v3.0.0...master)
+
+
+## [v3.0.0 (2023-10-16)](https://github.com/inspirum/xml-php/compare/v2.3.1...v3.0.0)
+### Changed
+- Support only **PHP 8.2+**
+- Make [`Config`](./src/Formatter/Config.php) interface instead of readonly class
+
+### Added
+- [`Node`](./src/Builder/Node.php) implements `\Inspirum\Arrayable\Model` interface
+- Added option cast node to flatten (one-dimensional) array
+- Added [`DefaultConfig`](./src/Formatter/DefaultConfig.php) config class
+- Added [`FullResponseConfig`](./src/Formatter/FullResponseConfig.php) config class
+- Added [`FlattenConfig`](./src/Formatter/FlattenConfig.php) config class
 
 
 ## [v2.3.1 (2023-08-08)](https://github.com/inspirum/xml-php/compare/v2.3.0...v2.3.1)